added query over tcp with ssl option
This definitely needs some work, just wanted a quick way to be able to
query over ssl.

When you call dns.resolver.query(), you can pass it an instance of
ssl.SSLContext and it will wrap the socket in it.
The function I added in dns.query is called `tcp_ssl`. Here are some
notes/todos about it:
- It uses blocking sockets
- It doesn't have options timeout, one_rr_per_rrset, and
ignore_trailing. Those probably should be added back in at some point.